chore(move-trigger): move trigger into separate function
